// src/app/api/swagger/models/api-market-deal-state-enum.ts
/* tslint:disable */
/* eslint-disable */
export enum ApiMarketDealStateEnum {
  WaitingOffer = 0,
  AgreementDeal = 1,
  AgreementContract = 2,
  CompletedDeal = 3,
  CanceledDeal = 4,
  ReadyForShipment = 5,
  Draft = 6,
  Retrading = 7,
  ReadyToPublish = 8,
  NotTookPlace = 9,
  Archive = 10,
  AuctionInProgress = 11
}